requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
ollama
python-socketio
aiofiles